Move segmenter options to options module
