+++ components/segmenter/tests/complex_word.rs
@@ -2,7 +2,7 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use icu_segmenter::{WordBreakInvariantOptions, WordSegmenter};
+use icu_segmenter::{options::WordBreakInvariantOptions, WordSegmenter};
 
 // Additional word segmenter tests with complex string.
