+++ components/segmenter/tests/css_line_break.rs
@@ -3,9 +3,9 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 use icu_locale_core::{langid, LanguageIdentifier};
-use icu_segmenter::LineBreakOptions;
-use icu_segmenter::LineBreakStrictness;
-use icu_segmenter::LineBreakWordOption;
+use icu_segmenter::options::LineBreakOptions;
+use icu_segmenter::options::LineBreakStrictness;
+use icu_segmenter::options::LineBreakWordOption;
 use icu_segmenter::LineSegmenter;
 
 fn check_with_options(
