+++ components/segmenter/tests/css_word_break.rs
@@ -2,9 +2,9 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use icu_segmenter::LineBreakOptions;
-use icu_segmenter::LineBreakStrictness;
-use icu_segmenter::LineBreakWordOption;
+use icu_segmenter::options::LineBreakOptions;
+use icu_segmenter::options::LineBreakStrictness;
+use icu_segmenter::options::LineBreakWordOption;
 use icu_segmenter::LineSegmenter;
 
 fn check_with_options(
